datebase bug fixed

+++ database/migrations/2021_06_02_072307_create_orders_table.php
@@ -16,12 +16,12 @@ class CreateOrdersTable extends Migration
         Schema::create('orders', function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->string('order_number')->unique();
-            $table->unsignedInteger('user_id');
+            $table->unsignedBigInteger('user_id');
             $table->foreign('user_id')->references('id')->on('users');
 
             $table->enum('status', ['pending', 'processing', 'completed', 'decline'])->default('pending');
             $table->decimal('grand_total', 20, 6);
-            $table->unsignedInteger('item_count');
+            $table->unsignedBigInteger('item_count');
 
             $table->boolean('payment_status')->default(1);
             $table->string('payment_method')->nullable();
